refactor: extract useWalletState hook from AppContext
Move 27 wallet/inventory/registry/drop/mint/whitelist useState calls + 11
callbacks into useWalletState. Accepts { setActionNotice, promptModal,
agentName, characterName } for cross-domain deps.

AppContext.tsx: 8,013 → 7,788 lines (-225)
useWalletState.ts: 397 lines (new)
Cumulative: 8,366 → 7,788 (-578 lines, 5 hooks extracted)
